pyproject.toml
[tool.poetry]
name = "pomodoroproject"
version = "0.1.0"
description = ""
authors = ["AliRN <ali_ramazanov_2000@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
uvicorn = {extras = ["standard"], version = "^0.32.0"}
flake8 = "^7.1.1"
pylint = "^3.3.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
ruff = "^0.7.2"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
